Feat: Added get request for /profile and post request for /update.
Get request for /profile occurs when the user clicks on 'Profile' button.
Users data are pulled from the database and saved to res.locals, profile.ejs rendered.
Post request for /update occurs when the users updates name or address.
The updated data are saved in the database, response is redirected to /profile with the new information.