Feat: Trade buttons are now functional.
The user can now request trades.
First it is checked to see if the user has any books to trade.
Then all user's books are displayed to choose one to trade.
Then all users owning the requested book are displayed so the user can choose which other user to trade with.
Finally when the request is placed, a message confirms to the user the successful submission.