Refactor: To send trade requests data to the client, post request on /trade and get request on /dashbaord have been modified.
Previously in a trade request, only books' isbns and other user's email were stored.
As a result, these information would have to be looked up in the database when dashboard is requested to list books' name and other user's name.
Now instead, these information are received from the client and saved directly in the inComingRequests and outgoingRequests.
This makes callback function for get request on /dashboard much smaller since it is not needed now to look up information in the shelf collection and other users.