Feat: Added a requestID field and post request on /cancel-request.
requestID is added to the request object and it consists of the string sum of the isbn of the two books.
The requestID is added to facilitate looking up the request in the database.
post request on /cancel-request removes the request object from both users and sends a 200 status response to the client.